erts: Use local trap exports for list functions
This prevents trapping from showing up in tracing, fixing certain
tests that assume the old and erroneous trace behavior for BIFs.

Note that this does not fix bogus arguments showing up on
exceptions as many of these functions still pass the remainder of
of the list to themselves when trapping.