Refactor win32 configure handling
There are a lot of MIXED_* options and adding a another mode for WSL
will make things worse, thus a cleanup is done first.

This commit consolidates the handling for the Makefiles, src and scripts and
moves the funcionality into similar named wrapper-scripts instead.
Path conversion is done with a script 'w32_path.sh' and
registry lookups with 'reg_query.sh'.

Path to the wrapper scripts is setup with: otp_build env_win32 <arch>.