v3_core: Fix wrapping of float/1 call
When translating guards, any calls to BIFs that are not known
to always return booleans should be wrapped in a comparison
with `true`. This was not done for `float/1` BIF (as a conversion
function, not the obsolete type test). Here is an example of
a function where the `float/1` calls were not wrapped:

    foo(X) when float(X) or float(X) -> ok.

Not wrapping `float/1` happens to be harmless, but is an
annyoing inconsistency.

With the correction, the guard will be rewritten like this
during the translatation to Core Erlang:

    foo(X) when (float(X) =:= true) or (float(X) =:= true) -> ok.