v3_kernel: Stop wrapping guards in try/catch
The old beam_bool pass (which was retired in 2016 in 348b5e6bee2f)
needed to have labels for all BIF calls in a guard. To force the
labels, v3_kernel wrapped all guards in try/catch block.

There is no longer any need to do this guard wrapping. Well, except
for a bug in v3_core that did not wrap calls to the unsafe type
test `is_function/2` in try/catch. With that bug fixed, the wrapping
can be removed.