v3_kernel: Optimize grouping of values in pattern matching
Use a map instead of a `dict`. A map is more efficient than
`dict`.

Furthermore, the `dict:append/3` operation has a quadratic
time complexity since it appends to a list. For some modules,
such as `Elixir.String.Unicode`, the lists can grow to thousands
of elements. With the optimization in this commit, the compilation
time spent in the v3_kernel pass for `Elixir.String.Unicode` was
reduced by about one third (down from about 1.5 seconds to less
than a second on my computer).