file_SUITE: Only test allocate/3 size updates on known ok platforms
Support is spotty in our test platform, especially on the rigs that
run the tests on funny file systems like NFS. We therefore limit
the test to the rigs where we know it will work under normal
circumstances.

That the call works at all is tested in allocate/1, this commit
only skips the size assertion test.