v3_kernel: Eliminate the no_usage annotation kludge
The variable in the `next` field in the #k_bin_seg{} and #k_bin_int{}
records is never used outside the matching code. That variable had
an `no_usage` annotation to prevent it from being included in the
set of used variables.

The `no_usage` annotation is not necessary. The existing mechanism
for calculation usage will work if we make sure that `pat_vars/1`
include the variable in `next` fields in the set of new variables.