ssl: Improve session handling
Due to legacy reasons cleaning of the session handling table was
complex and needed to be delayed. However the implementation
evolved many times since and this commit removes the need for the
delay. Hopefully this will also remove the problem with server
session tables that grow quicker than they can shrink under a
heavy load.

This also fixes TLS-1.3 compatibility with pre TLS-1.3 sessions.