Correct the documentation for proplists:substitute_negations/2
According to the documentation, the following call:

    proplists:substitute_negations([{no_foo, foo}], [{no_foo, false}]).

should return `[{foo, true}]`. The actual return value is `[foo]`.

The implementation of `proplists:substitute_negations/2` calls
`proplists:property/2` to reduce the substituted items to their
minimal representation. That seems to be the intention of the
author and it makes sense.

Therefore, the documentation should be corrected.

While at it, add a minimal `proplists_SUITE` that tests all examples
in the documentation.