beam_ssa_type: Infer return types from given arguments
We can narrow down return types by keeping track of the function
argument types at each return.

When determining the return type of a function, we only consider
the returns where the given function arguments can be met with the
stored ones, as the return can only be reached when the types are
compatible.