ets_hash_decentralized_counters: change default option
This commit changes the default value for the ETS table option
decentralized_counters. The default value for decentralized_counters
in ETS tables of type ordered_set with the write_concurrency option
enabled is true and the default value for decentralized_counters in
all other tables is false. This commit also changes the ETS
documentation accordingly.