Repeat optimization passes until a fixpoint is reached
The compiler would repeat some optimization passes a fixed number of
times. For example, the type optimization pass would be run three
times on all functions in a module. That could be too much for some
functions (wasting compilation time) or too little (missing some
possible optimizations).

This commit updates `beam_ssa_opt` to run the repeated passes to
a fixpoint. Only functions that could potentially be improved by
additional optimizations will be optimized again.

Thus, the type optimization pass is run twice for all functions.
It will be repeated for functions that would benefit from it
because better type information is available or because its code
was changed in the previous iteration.