v3_kernel: Extend the capabilities of #c_letrec{}
This is a first in a series of commits to allow pattern matching to be
done, partially or wholly, in Core Erlang without duplicating code.

Here is an example of the kind of transformation we want to be able
to do. Here is the first an example in Erlang:

    foo(X) ->
        case X of
            <<Size:32,Bin:Size/binary>> ->
                {ok,Bin};
            _ ->
                error
        end.

The example can be translated to Core Erlang like this:

    'foo'/1 =
        fun (X) ->
            case X of
              <#{#<Size>(32,1,'integer',['unsigned'|['big']]),
                 #<Bin>(Size,8,'binary',['unsigned'|['big']])}#> when 'true' ->
                  {'ok',Bin}
              Other when 'true' ->
                  'error'
            end

Here the `Size` variable is first bound to a value and then used in in
the same pattern. Having to support that a variable can be both new
and used in the same pattern complicates all compiler passes that have
to deal with it. We would like to rewrite the example like this:

    'foo'/1 =
        fun (X) ->
            case X of
              <#{#<Size>(32,1,'integer',['unsigned'|['big']]),
                 #<Tail>('all',8,'binary',['unsigned'|['big']])}#> when 'true' ->
                  case Tail of
                    <#{#<Bin>(Size,8,'binary',['unsigned'|['big']])}#> when 'true' ->
                        {'ok',Bin}
                    Other when 'true' ->
                        'error'
                  end
              Other when 'true' ->
                  'error'
            end

This works, but note that the body of the second clause (the atom `error`)
has been duplicated.

This commit makes it possible for v3_kernel to translate letrecs,
making it possible to support gotos in Core Erlang. Here is the
example rewritten using a `letrec`, eliminating the code duplication:

    'foo'/1 =
        fun (X) ->
        ( letrec
              'second_clause'/0 =
                  fun () ->
                      'error'
          in  case X of
                <#{#<Size>(32,1,'integer',['unsigned'|['big']]),
                   #<Tail>('all',1,'binary',['unsigned'|['big']])}#> when 'true' ->
                    case Tail of
                      <#{#<Bin>(Size,8,'binary',['unsigned'|['big']])}#> when 'true' ->
                          {'ok',Bin}
                      Other when 'true' ->
                            apply 'second_clause'/0()
                    end
                Other when 'true' ->
                      apply 'second_clause'/0()
              end
          -| ['letrec_goto'] )

The `letrec_goto` annotation added to the `letrec` signals to
`v3_kernel` that no new function should be created. The `apply
'second_clause'/0()` calls will be translated to unconditional
branches to the the label corresponding to `second_clause` in
the SSA code.