erts: Honor CPU quotas when deciding number of online schedulers
Using more cores than we have quota for is likely to hurt
performance as any busy-waiting we might do (waiting for new jobs
to arrive in a queue, waiting for a lock, etc) will burn the quota
without doing useful work, leaving us with fewer resources to work
with. This is especially bad when the +sbwt option has been used to
increase the wait time.

This commit checks our current CPU quota (if possible) and tries to
limit the number of online schedulers accordingly on startup. The
schedulers can be brought online later on, or overridden with the
`+S` option.