[esock|test] Add test case for the 'pktinfo' ip option (UDP)
Add a simple testcase for the 'pktinfo' socket 'ip' option (UDP).

Note that its should be possible to send pktinfo also (that
is the sender provides pktinfo in the control message header
part of the message) when using sendmsg, but this have not
yet been implemented (it *could* be sent in raw form), so we
do not test that part.

OTP-16114