v3_kernel: Simplify translation of function_clause failures
Translation of the `match_fail` primop for a `function_clause` failure
is complicated because the invocation of `match_fail` might have been
inlined into a different function. Since a `function_clause` exception
would be confusing, the `function_clause` exception is replaced with a
`case_clause` exception.

In order to know whether a `match_fail` primop has been inlined,
`v3_core` would add a `function_name` annotation to the `match_fail`
primop, and the inliner would remove the annotation.

This commit simplifies the translation of `function_clause` by
comparing the arguments in the `function_clause` with the arguments of
the function. If the are identical (in order and number) a
`function_clause` exception will be emitted, otherwise a
`case_clause`.