Erlang emacs-mode: Store URL together with downloaded man pages
The commit makes sure that the download URL gets stored in a file
after man pages have been downloaded by the erlang-mode. The
interactive Emacs lisp function erlang-man-download will check the URL
that is stored in this file and re-download the man pages if the URL
has changed (i.e., the value of the variable erlang-man-download-url
has changed or the user has typed another URL after a call to
erlang-man-download-ask). Thanks to @johanclaesson (Johan Claesson)
for suggesting this functionality.

This commit also fixes a bug that caused the man menu items under the
Erlang Emacs menu to not get displayed when the erlang-root-dir
variable was nil even when man pages existed in the cache.