[esock|test] Add proper peek-off test case
Add a proper test case for the peek-off 'socket' option.

This option is dangerous at the moment. If you do a second
socket:recv(Sock, 0, [peek]), the call will hang (unless a
timeout is specified)!!
The reason for this is that we use a select to detect if
there is something to read, but when the second call is made,
the peek pointer has been moved to the end of data, so the
second peek read will fail with eagain even though there is
data to read. The code will then do a new select. Wich will
trigger immediately (since there is data to read). Ando so on.

So, either disable this option at the moment, or add some
code that keeps track on the fact that we have changed
peek-off (and can break the loop).

OTP-16114
OTP-16196