Add info about compiler latin1 support without encoding string removal
This commit adds information to the warning message that appears when
the compiler detects a non-utf-8 encoded source files without an
encoding string in the beginning of the file. Now, the message
contains information that support for latin1 encoded source file
without an encoding string will be removed in Erlang/OTP 24.