Refactor handling of matched out sizes in binaries
A long time ago binary patterns were extended to allow using a matched
out variable as a size of another segment. For example:

    <<Size:32,Data:Size/binary>> = Binary

This extension was carried over to Core Erlang, which made it much
more complicated to correctly optimize Core Erlang, since a variable
in a pattern could be both new and used.

This commit rewrites matching of binaries to split up patterns that
match out a variable and uses it in the same pattern. Here is an
example:

    foo(X) ->
        case X of
            <<Size:32,Bin:Size/binary>> ->
                {ok,Bin};
            _ ->
                error
        end.

This function will now be translated to Core Erlang like this:

    'foo'/1 =
        fun (X) ->
        ( letrec
              'second_clause'/0 =
                  fun () ->
                      'error'
          in  case X of
                <#{#<Size>(32,1,'integer',['unsigned'|['big']]),
                   #<Tail>('all',1,'binary',['unsigned'|['big']])}#> when 'true' ->
                    case Tail of
                      <#{#<Bin>(Size,8,'binary',['unsigned'|['big']])}#> when 'true' ->
                          {'ok',Bin}
                      Other when 'true' ->
                            apply 'second_clause'/0()
                    end
                Other when 'true' ->
                      apply 'second_clause'/0()
              end
          -| ['letrec_goto'] )

Note that when compiling using `from_core`, it is an error to bind
and use a variable in the same pattern. Compilers that generate
Core Erlang code must split such patterns into nested cases.

The mechanism added in this commit to split clauses should also be
possible to use to handle, for example, expressions in binary
patterns. Of the compiler passes after the front end, only `v3_core`
would have to worry about handling the size expressions.

One potential issue with the new way of translating to Core Erlang
is that the pattern matching compiler in `v3_kernel` does not see
all clauses at once and may generate worse code, that is, evaluating
the clauses sequentially instead of in parallel. Here is an example:

    bar(a) -> a;
    bar(<<Size:32,Bin:Size/binary>>) -> Bin;
    bar(c) -> c.

Here `v3_kernel` will first generate code for the first two clauses
and then for the last clause. It will not be able to combine the
matching of the two atom clauses to a single select instruction.

However, in this simple case, the SSA code optimizations will be able
combine the matching of the two atoms to a select_val instruction.

We will probably have to make the pattern matching compiler smarter
at handling nested cases to avoid producing worse code than in
OTP 22. Alternatively we could add additional optimization of
the SSA code.