erts: Fix sigq deferred_save when inner queue is empty
If the inner queue is empty, the deferred_save would incorrectly
trigger and set the message queue pointer to the wrong place.
This commit makes it so that the message queue optimization does
not trigger when the inner queue is empty. This is a correct
solution, but not ideal as it means that the message queue optimization
will not help when you have a huge middle queue but en empty
inner queue.