[net] Remove use of some (eventually) deprecated idn flags
In (at this time) new versions of gcc the flags NI_IDN_ALLOW_UNASSIGNED
and NI_IDN_USE_STD3_ASCII_RULES are deprecated. That is, they exist
but results in a compiler (deprecated) warning when used. And since we
are trying to have code that does not generate warnings, the simplest
solution is to not use these flags.
Also, fixed a flag spelling (was nomericserv, should have been numericserv).

OTP-16203