erts: Optimize deferred save message queue pointer
In 765fa3285ef a bug fix for the message queue receive
optimization made the optimization not take when the
internal message queue was empty and the save mark was
in the middle queue.

With this commit we move that decision into a flags instead
so we now can know when to restore the save pointer when
messages are moved from the middle to internal queue.