Do not merge maps when they are the same or empty
If merging two maps that are precisely the same
in memory or one of them is empty, we can avoid
the whole merge operation and simply return the
given map.

This can reduce memory allocation and CPU work for
algorithms that may be updating a map and merging
into a previous value. Or when providing defaults
which may be empty. The check is also very cheap,
which means the downsides are minimal.

The commit also adds coverage for a similar behaviour
already implemented for maps:put/3.