ets: Remove table locking when using smp 1
If we only have one normal scheduler we know that only
one thread can access an ETS table at a time, so we don't
need to optimize away that.

This approach reads and checks the global variable
erts_no_schedulers each time, it could be optimized
to use the table flags field instead. Though I'm not sure
how much that would help.

With this change doing ets:lookup/2 och a small value is
about twice as fast.