Optimize empty 'receive' with only an 'after' clause
A `receive` with only an `after` clause is currently optimized in
`beam_a` to remove the redundant `loop_rec` and `loop_rec_end`
instructions. Optimizing that late is a little bit awkward because
`beam_ssa_recv` will try to optimize the `receive` by adding
`recv_mark` and `recv_set` instructions, which `beam_a` will have to
remove. It can also result in a larger stack frame than necessary
because the `peek_message` instruction will force the timeout value to
be saved to the stack frame.

Modify `beam_kernel_to_ssa` to never generate the `peek_message`
and `recv_next` instructions for a `receive` with no clauses, and
also teach the SSA optimization passes to cope with it.