dialyzer: Fix an overloaded contract warning
When creating the missing_range warning, the supremum of all function
type ranges is considered, not the range of one function type at a
time.

Notice that more extra_range messages than before can be emitted (the
check for extra_range was not always performed if a missing_range was
found).