ssl: Implement manual ticket handling
This commit implements mechanism for manual handling and configuration
of session tickets in TLS 1.3 clients.

When option 'session_tickets' is set to enabled in a TLS 1.3 client,
it will not store the tickets sent by the server in the ticket store
but it will deliver it as a message to the user application:

{ssl, session_ticket, {SNI, TicketData]}}

TicketData is an opaque binary that can be used in the 'use_ticket'
option that expects a list. SNI is the Servíce Name Indication.

Overhaul ssl_session_ticket_SUITE and verify manual and automatic
ticket handling.