[kernel] Attempt to resolve make awk warning
When building kernal from its top directory, an
awk warning is generated (cd lib/kernel && make)
from make in the doc/src directory even though
nothing is actually built there.
The root cause is the target 'xmllint' and the
variables that are evaluated (in the otp_release_target.mk),
containing all xml files in kernel doc. Some of which
are *generated* (at a later stage) and therefor does
not yet exist...

Trying to always include the net documentation to
avoid the above generation.

OTP-16294