v3_kernel: Eliminate #k_protected{} and #k_protected_value{}
The #k_protected{} and #k_protected_value{} constructs are
restricted variants of try/catch used in guards. By using
the #k_try{} record instead, v3_kernel can be somewhat simplified.

This refactoring also makes it possible to use expressions
protected in the same way as in guards in the body of the function.
That would be potentially useful in the future.