Allow changing the -mode when the system restarts
In order for Elixir developers to configure their
systems using Elixir syntax, the Elixir application
must be started. However, in order to start the
Elixir application, kernel and stdlib need to be
started, which means we can't configure those
applications.

Elixir addresses this by booting twice. First, the
system boots with a minimal set of apps started, then
it executes the configuration files, writes them to
disk and restarts the system, which will pick up the
new configuration.

The problemw with such approach is that, when running
in embedded mode, starting and shutting down the system
takes a long period of time, due to the loading and
purging of all modules.

This PR addresses the problem by allowing the mode to
be changed to embedded on init:restart/1. This allows
Elixir to first boot in interactive mode, load the
configuration, and then start the system in embedded
mode.

This reduces the boot time in the sample application
that reproduces the problem from 5s to 1s.