compiler: Make map types more precise
While we can't track individual elements (further explained in
?TUPLE_ELEMENT_LIMIT), we can track the general key and element
types, which is often sufficient since maps tend to be uniform.

The effect in OTP is rather limited given the rarity of raw maps
that never escape the module, but it should be greatly amplified
if we get cross-module optimization to work.