erts: Include block types in allocation statistics
Now that carriers can migrate freely, a binary allocator may find
itself responsible for a bunch of heap fragments. We used to
present such blocks as if they were still in the pool ("foreign
blocks") but it was a bit too counter-intuitive.

This commit presents all blocks on the allocator that employs their
carrier, regardless of their type.