logger: Make sure to not recursively crash for internal log events
If a formatter crashes, this will generate an internal log event that
will be sent to all loggers. If the logger then again crashes this will
create a new log event and so on.

This fix adds a meta-data tag that makes it so that the internal log events
don't continue until infinity but instead stop and are only logger by the
simple logger without any formatters in the end. This so that we can get some
information about what has gone wrong.