Add tag to response message
Since users may now inspect the message queue between,
'send_req' and 'wait_response' which may contain the reply message
which is a bit non-descriptive, a tag is added to explain it.

This doesn't mean that the message interface is set in stone or
documented, it can still be changed.