[snmp|test] Add utility function for host info
Added a utility function to "analyze" the host
and print various info about it; OS, CPU and
Memory, if any of these things can be found
(they can on linux, FreeBSD and OpenBSD).
Also make a rough calculation pf how powerful
the host is and present it as a 'factor' in
the config (snmp_factor').
This function is called in the common init_per_suite
function that all snmp test suite call.