[ERL-588] Add mnesia funcs to TOC
Prior to this commit mnesia functions were not added to the HTML Table
of Contents because the doc generation wasn't pulling the specs from erl
files. This commit adds a top-level docs specs file to mnesia and the
appropriate definitions in the Makefile in order for the specs to be
extracted during `make docs`.