[megaco|test] Improved (slave) node handling
A test case failed because a node died before the
test started. That is, the start_nodes function
did not start the node because it was already running,
but then it died. So when the test case started,
MGC start failed because the node it was to run on
did not exit...

So, instead of assuming that node *could* exist
(and "reuse" it), make it possible to require a
freshly started node.
This happened in the pending-limit suite, so the
changes has been done there (and in the lib module).