remove docs for mnesia:dirty_slot/2
Prior to this commit, the mnesia documentation was generating entries
for mnesia:dirty_slot/2; however, this function does not have a spec and
additionally has a comment that it should not be used and is only there
for backward compatibility. This commit therefore removes the
documentation of this function to reduce future use.