export uri_string:decode/1
In the current state, the uri_string module parses URIs to maps of components
but does not decode those which can be percent-encoded such as the
fragment. It is not a problem with the query since `uri_string:dissect_query/1`
will decode it, but other components need decoding.

Exporting `uri_string:decode/1` solves this problem and lets users decode
components when they need it.

In the future, it could be nice to add `uri_string:parse/2`, the second
argument being a list of options which could include `decode_components`.