Optimize matching of transformations
The transformation engine is rather naive in that it runs in sequence
all transformations that starts with a given instruction. This commit
will make the engine slightly smarter so that it can quickly skip a
series of transformations that can't possibly match.

There are currently 49 transformation rules that match a sequence of
instructions that begin with a `move` instruction. Some of the rules
matches very seldom. Therefore it is important that the engine can
quickly skip transformations that can't possibly match.