ssl: Correct DTLS listen emulation
When a DTLS "listen socket" is closed the owner shall be
deassociated with the packet multiplexor process if it should be keept alive.
The packet multiplexor may still have ongoing connections. When a new
listen socket is opened for the same port it shall be associated with the
existing packet multiplexor if it is still alive.