Optimize relational tests
Optimize the `is_lt`, `is_ge`, `is_eq`, and `is_ne` instructions (used
to implement the Erlang operators `<`, `=<`, `>=`, `>`, `==`, and
`!=`).

The main optimization comes from the observation that when one of the
operands for `is_lt` or `is_ge` is a literal, that literal is almost
always an integer and almost never an atom. Therefore, it is a waste
of time and code space to compare atoms before calling the general
comparison function.

While at it, also slightly optimize the `is_eq` and `is_ne` instructions.