Fix incorrect transport options
Commit 5f3becad caused transport options differing from those passed to
diameter:add_transport/2 to be used in several situations: when starting
a transport process after connect_timer expiry after an initial
connection attempt has failed, when starting a transport process after a
connection has been accepted, when sending events, when returning
options in diameter:service_info/2, and possibly more.

In particular, the configuration options below were lost.

  avp_dictionaries
  incoming_maxlen
  spawn_opt
  strict_mbit

The effect was as if these hadn't been configured.

Discovered when testing {diameter_dist, route_session, []} as a
spawn_opt MFA and seeing that the callback was missing after an initial
connection failure. (Thanks to Tamas Bodnar.)

Add a simply match to the event suite that detects one instance of the
fault.