[megaco|test] Tweaking the trans test suite
Mostly adding and improving printouts to be able to see
what is actually sent and received (in the megaco generators).
The test case that is improved is multi_trans_req_and_ack_reqmaxcount,
which has failed on a windows host.
Also adding info/warning/error messages to the await-completion
functions used to wait for generator(s) completion (to make it
easier to see shich generator actually fails).