erts: Suppress warning "'ycf_stack_alloc' defined but not used"
by introducing macro YCF_GCC_ATTRIBUTE_UNUSED.

The YCF_GCC_DIAG_OFF/ON(unused-function) did not seem to work.
My guess is you would have to keep it OFF for the rest of the file
as it's at the end of the file gcc discovers that a function is
actually unused.