erl_lint: Fix forbidden segments in binary generators
In the pattern for bit string generators, segments without size are
not allowed. Thus erl_lint correctly generates an error for the
following bit string generator:

    foo(Bin) ->
	[ [X] || <<X/bitstring>> <= Bin ].

However, if the pattern matched a literal erl_lint would not generate
an error:

    bar(Bin) ->
	[ 42 || <<42/bitstring>> <= Bin ].

For this particular example, the compiler would crash later. For
variations of this code, the compiler would not crash but would
generate nonsensical code.