In binary matching, make explicit use of `all` illegal
The literal atom `all` is used in the final segment of a binary
pattern to indicate that all of the remaining binary should match.
However, `all` cannot be explicitly, but is added by v3_core when
no size has been specified.

Replace any explicit use of the `all` atom by the atom `bad_size`.