beam_ssa_opt: Be more careful in ssa_opt_try optimization
The ssa_opt_try optimization can discard instructions in a block that
ends in a `kill_try_tag` instruction. This was not noticed because
blocks that contain `kill_try_tag` usually don't contain other
important instructions. However, if optimization are selectively
disabled significant instructions can end up in the same block as
`kill_try_tag`. For example, if the beam_ssa_bool pass is disabled,
the compiler will crash because a phi node is placed at the beginnig
of a block with a `kill_try_tag` instruction:

    in_interval(Value, {Rlow,Rhigh}) ->
	try
	    Rlow =< round(Value) andalso round(Value) =< Rhigh
	catch
	    _:_ -> false
	end.

Make the ssa_opt_try optimization more robust by preserving important
instructions in blocks that have a `kill_try_tag` instruction.