[megaco] Add (sdp) wrapper function to interface module
An already documented (wrapper) function was acually missing.
The function megaco:get_sdp_record_from_PropertyGroup/2 was
documented but did not exist in the main interface module
(megaco.erl). It did exist inte the sdp module (megaco_sdp.erl).
So, instead of documenting the megaco_sdp module, we added
the missing wrapper function in the megaco module (so that
the documentation is, finally, correct).

OTP-16449