Fix traffic counters for remote request handler processes
Commit f1cdd721 documented the possibility of starting a handler process
for an incoming Diameter request with a configured MFA, but the counter
implementation (ie. {traffic_counters, true}) was broken when this
process was not on the same node as the transport configuration:
incrementing a counter was a noop since the transport pid was not
associated with the transport reference in diameter_stats on the remote
node.

This was a good thing though, since distributing the counters across
handler nodes has a number of problems: not losing counters if the
connection to a handler node is lost, or not being able to remove
counters on a handler node when the transport is removed, keeping track
of where counters are located if they have to be retrieved from multiple
nodes, and so on. Moreover, it's the service/transport configuration on
the local node that orders the counters, so incrementing them on remote
nodes is questionable.

Avoid the problems by incrementing counters where the transport in
question is configured, the node that terminates any associated peer
connection. Instead of calling diameter_stats to increment the counter
when the diameter_peer_fsm process it's being incremented for is not
local, bang the counter to that process to let it call diameter_stats.
The only requirement on the handler node is now that diameter and
application callbacks reside on the code path, and the documentation has
been updated to reflect this.

Compare this to the case of outgoing requests, where counters are
incremented on the originating node. Similarly, it's the local
configuration that orders the counters, and there can be multiple
remote nodes through which the request can be sent on a peer connection
just like there can be multiple handler nodes in the case of an incoming
request.