[megaco|test] More 'trans' test suite tweaking
A test case in the 'trans' suite misbehaves.
This test case has to do with the transaction sender process.
The timeout for it is set to 5000.
We get a reply for one of the requests that is suppose to be sent
from the transaction sender, but is unknown (by the MG) when it is
received. So added printout of which requests we know about after
each request send and reply receive.
Also include time (and process) in the printouts, so that we can
see more info (about the printer).