sys_core_fold: Remove transformation meant to help the code generator
sys_core_fold would attempt to move a safe boolean expression in a
`case` into a guard. The purpose of the transformation was to help
v3_codegen (the code generator before OTP 22) to generate better code.

With the introduction of the SSA compiler passes, this help is
no longer needed.