Skip many more traffic testcases
There are currently 55 traffic testcases and 1440 config variants:
55*1440 = 79200 possible testcases. Skipping 25% of these (randomly)
still leaves 59400 for common_test to trudge through, and this is not
quick, even though some of these 54900 are also skipped: almost 90
minutes on one reasonable host.

The benefit of running all these variants of the same test ad nauseum
(on many hosts to boot) is doubtful, so increase the number skipped to
90%. This leaves 7920 testcases, still not a small number, and takes
around 12 minutes vs 90. (Which is exactly 90 * 10/75 in fact.)