Skip traffic tests differently
The reduction of the number of traffic cases in the parent commit speeds
up testing, but common_test doesn't seem to do well with the large
number of groups returned from groups/0: one for each of the 1440 config
variants and a few more.

Randomly skip variants when building the groups list instead of when
init_per_group/2 is invoked, to reduce the running time from 12 minutes
mentioned in the parent commit to around 8.